requirements.txt
Django==5.2.13
djangorestframework==3.15.2
psycopg2-binary==2.9.9

# ASGI
channels==4.0.0
daphne==4.0.0
channels_redis==4.1.0

# Auth
django-allauth==0.63.3
djangorestframework-simplejwt==5.3.1

# DB & env
dj-database-url==2.2.0
python-dotenv==1.0.1

# Redis
django-redis==5.4.0

# CORS
django-cors-headers==4.3.1

# Editors (keep only ONE if possible)
django-ckeditor-5==0.2.12
django-summernote==0.8.20.0
django-tinymce==4.1.0

# Utilities
pillow==10.3.0
requests==2.31.0

# AWS (only if using S3)
boto3==1.34.162
botocore==1.34.162

# Production
gunicorn==21.2.0
whitenoise==6.6.0
